requirements.txt
spotipy
pathlib
python-dotenv